+++ apps/server/src/modules/learnroom/uc/dashboard.uc.spec.ts
@@ -52,7 +52,10 @@ describe('dashboard uc', () => {
 					grid: [
 						{
 							pos: { x: 1, y: 2 },
-							gridElement: new GridElement('elementId', new DefaultGridReference('referenceId', 'Mathe')),
+							gridElement: GridElement.FromSingleReference(
+								'elementId',
+								new DefaultGridReference('referenceId', 'Mathe')
+							),
 						},
 					],
 				});
@@ -71,7 +74,10 @@ describe('dashboard uc', () => {
 							grid: [
 								{
 									pos: { x: 1, y: 2 },
-									gridElement: new GridElement('elementId', new DefaultGridReference('referenceId', 'Mathe')),
+									gridElement: GridElement.FromSingleReference(
+										'elementId',
+										new DefaultGridReference('referenceId', 'Mathe')
+									),
 								},
 							],
 						})